requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
xgboost>=1.5.0
mlflow>=2.0.0
opencv-python>=4.5.0
pycocotools>=2.0.2
matplotlib>=3.4.0
PyYAML>=5.4.0
joblib>=1.0.0
torch>=1.10.0
